requirements.txt
requests~=2.31.0
python-decouple==3.6
python-telegram-bot~=20.5
python-telegram-bot[job-queue]
APScheduler~=3.10.4
pymongo~=4.5.0